Fix minio URL from core service
The URL used by the core service to contact the minio service was identical for
the production and staging services. This led to a conflict when performing the
setup and during operation.

The fix causes the prod/staging environment to use the container name instead
of the service name, which is the same for both instances. The container name
has the format of 'sdg-server-$DEPLOY_TYPE-minio'. This also required passing
the deploy type as a variable into the core service.